Add UsePLT, fix some bugs

+++ include/pass/SplitBlocks.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <cstddef>
+
 namespace LL2X {
 	class Function;
 }
@@ -7,5 +9,5 @@ namespace LL2X {
 namespace LL2X::Passes {
 	/** Splits basic blocks such that no basic block contains more definitions than the number of physical registers.
 	 *  Returns the number of times a block was split. */
-	int splitBlocks(Function &);
+	size_t splitBlocks(Function &);
 }
